Move renderDashboard to component file
