+++ packages/dashboard/package.json
@@ -50,9 +50,12 @@
     "@types/cypress-image-snapshot": "3.1.6",
     "cypress": "^7.7.0",
     "cypress-image-snapshot": "4.0.1",
+    "@vue/cli-plugin-typescript": "^4.5.15",
     "@vue/cli-service": "^4.5.15",
+    "@vue/compiler-dom": "^3.2.26",
     "cypress-wait-until": "^1.7.2",
     "postcss-url": "^10.1.3",
+    "vue": "^3.2.26",
     "rollup-plugin-node-polyfills": "^0.2.1"
   },
   "dependencies": {
